+++ neqo-http3/src/lib.rs
@@ -6,8 +6,12 @@
 
 #![deny(warnings)]
 
+mod client_events;
 pub mod connection;
+mod control_stream_local;
+mod control_stream_remote;
 pub mod hframe;
+mod stream_type_reader;
 mod transaction_client;
 pub mod transaction_server;
 
@@ -17,7 +21,8 @@ pub use neqo_transport::Output;
 
 use self::hframe::HFrameType;
 
-pub use connection::{Http3Connection, Http3Event, Http3State};
+pub use client_events::Http3ClientEvent;
+pub use connection::{Http3Connection, Http3State};
 pub use neqo_qpack::Header;
 pub use transaction_server::TransactionServer;
